Filters instagram photos by last 30 days
Was having an async issue with trying to make my ajax call multiple
times, so i turned async off.  Also Instagram was only giving me 20
results at a time so that is the reason for multiple ajax calls.

Next I will be working on ranking pictures based on number of comments,
if comments are equal then i’ll return the newest post.