Posts flickr pictures from the past month
The initial page load is a tad bit longer than what I want it to be.
Also the React code is becoming a little bit gross looking.

I want to refactor and be able to speed up the initial load time.  Will
do research and try to implement.

Also I need a piece that lets people know which pictures are rank the
highest by comments.